Import file supports html originating from external songbooks.
