+++ editor/conv/lyric.test.js
@@ -1,5 +1,5 @@
 import { JSDOM}  from 'jsdom';
-import { interpretationContent2lyric } from './converter.js';
+import { interpretationContent2lyric } from './conv.js';
 
 function ic2l(html) {
     const dom = new JSDOM();
